docs: Update repository documentation and structure
- Enhance README.md with comprehensive repository structure and usage guidelines
- Add detailed descriptions for all scripts in Scripts/README.md
- Include legal and ethical considerations section
- Add contributing guidelines and license information
- Remove listener.py as it's now included in Scripts directory
- Improve organization with emoji indicators and clear categorization
